requirements.txt
tensorflow==2.15.0
opencv-python
pandas
matplotlib
scikit-learn
numpy
pillow==10.4.0